A RUNTIME FONT PARSER FOR AS3
here is all the files that I used whilst porting a java library for runtime fonts into actionscript 3. It is working and I am now making it an open source project to help further its development. In time we will trim this down to just being the source files but on this first commit I wanted the full history of the port to be available. So check out this first revision if you want the full history of the tool as future commits will see deletion of a lot of these archive files as we move to just provide a more complete product.

thanks

mike