LIGHTER CHECKOUT
Im just deleting project resources to make the checkout of this project lighter.

so this version doesn't have all the project history files and resources for the original port to current state. please check this out as its more lightweight... if you want the full port history and all resources and versions check out the first revision of this project.